+++ NEWS.md
@@ -1,3 +1,5 @@
+# predictNMB (development version)
+
 # predictNMB 0.2.0
 
 * Track QALYs and intervention costs