+++ NEWS.md
@@ -1,4 +1,4 @@
-# predictNMB (development version)
+# predictNMB 0.2.1
 
 * `ce_plot()`